Added back app delegate marks, moved the flex debugger setup to the AppCoordinator.
